// web/legacy_web_app/src/lib/components/Backgrounds/types/types.ts
export type Dimensions = {
  width: number;
  height: number;
};
// ...
export interface SantaState {
  x: number;
  y: number;
  speed: number;
  active: boolean;
  direction: number;
  opacity: number;
  imageLoaded?: boolean;
}